+++ mitra_activitypub/src/builders/add_note.rs
@@ -70,7 +70,7 @@ pub(super) async fn get_add_note_recipients(
     let mut recipients = vec![];
     for profile in followers {
         if let Some(remote_actor) = profile.actor_json {
-            recipients.extend(Recipient::from_actor_data(&remote_actor));
+            recipients.extend(Recipient::for_inbox(&remote_actor));
         };
     };
     Ok(recipients)