+++ mitra_activitypub/src/builders/undo_follow.rs
@@ -83,7 +83,7 @@ pub fn prepare_undo_follow(
         follow_request_id,
         follow_request_has_deprecated_ap_id,
     );
-    let recipients = Recipient::from_actor_data(target_actor);
+    let recipients = Recipient::for_inbox(target_actor);
     Ok(OutgoingActivityJobData::new(
         &instance.url(),
         sender,